fix(arch): resolve dependency violation — service layer must not import infrastructure
- Restore github_issue_creator interface with CreateIssueInput/CreatedIssue types
- Service layer depends on interface, not concrete githubclient.Client
- githubclient implements GitHubIssueCreator interface (compile-time check)
- DI wiring passes concrete client through interface
